+++ NEWS.md
@@ -1,3 +1,7 @@
+# BioDataScience 2024.1.0
+
+-   Help addin recording.
+
 # BioDataScience 2024.0.0
 
 -   New version for academic year 2024-2024. New passwords.